internal/buildctx/google-docs-mcp/pyproject.toml
[project]
name = "mcp-google-docs-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.122.0",
    "fastmcp>=2.13.1",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "google-auth-oauthlib>=1.2.2",
    "pydantic-settings>=2.6.1",
    "markdown-it-py>=3.0.0",
]
